requirements.txt
# Nexus Miracle - Dependencies
# Python 3.11+ required

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1

# HTTP Client
httpx==0.27.0

# AI Services
google-generativeai==0.8.0
elevenlabs==1.0.0

# Voice Activity Detection
silero-vad==5.1

# Configuration & Utilities
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# JSON Performance
orjson==3.10.0

# Audio Processing
numpy>=1.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0
